# gen/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Grand Kapital
  description: Взаимодействие с HTTP-сервисом, эмулирующий банковские транзакции и менеджмент пользовательский данных
...
paths:
...
  /emails:
    post:
      tags:
        - emails
      summary: Add email to user
      operationId: addUserEmail
      description: Add new email address for user (must be unique)
      parameters:
        - name: userId
          in: query
          description: user id
          required: true
          schema:
            type: integer
            format: int64
        - name: email
          in: query
          description: new email
          required: true
          schema:
            type: string
      responses:
        200:
          description: Email added successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Email'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - emails
      summary: Remove email from user
      operationId: removeUserEmail
      description: Remove email address from user profile
      parameters:
        - name: userId
          in: query
          description: user id
          required: true
          schema:
            type: integer
            format: int64
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Email removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - emails
      summary: Replace email with new one
      operationId: changeUserEmail
      description: Replace user email address
      parameters:
        - name: userId
          in: query
          description: user id
          required: true
          schema:
            type: integer
            format: int64
        - name: oldEmail
          in: query
          required: true
          schema:
            type: string
        - name: newEmail
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Email replaced successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /phones:
    post:
      tags:
        - phones
      summary: Add phone to user
      operationId: addUserPhone
      description: Add new phone number for user (must be unique and in format 79200234567)
      parameters:
        - name: userId
          in: query
          description: user id
          required: true
          schema:
            type: integer
            format: int64
        - name: phone
          in: query
          description: new phone
          required: true
          schema:
            type: string
            pattern: '^7\d{10}$'
      responses:
        200:
          description: Phone added successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Phone'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - phones
      summary: Remove user phone
      operationId: removeUserPhone
      description: Remove user phone number
      parameters:
        - name: userId
          in: query
          description: user id
          required: true
          schema:
            type: integer
            format: int64
        - name: phone
          in: query
          required: true
          schema:
            type: string
            pattern: '^7\d{10}$'
      responses:
        200:
          description: Phone removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - phones
      summary: Replace phone with new one
      operationId: changeUserPhone
      description: Replace user phone
      parameters:
        - name: userId
          in: query
          description: user id
          required: true
          schema:
            type: integer
            format: int64
        - name: oldPhone
          in: query
          required: true
          schema:
            type: string
            pattern: '^7\d{10}$'
        - name: newPhone
          in: query
          required: true
          schema:
            type: string
            pattern: '^7\d{10}$'
      responses:
        200:
          description: Phone number change successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Email:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          example: "new.email@example.com"
    Phone:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          pattern: '^7\d{10}$'
          example: "79207654321"
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          description: "message of error"
          example: "Invalid date format. Expected DD.MM.YYYY"
      required:
        - message
    Message:
      type: object
      properties:
        message:
          type: string
          description: "success message"
          example: "Success!"
      required:
        - message
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT